Implement alpha blending ourselves before passing it on to printpdf
This ensures alpha is actually blended without generating an sbox, as
the sbox is rendered incorrectly by some pdf viewers.

Previously, the image was implicitly converted to an sbox by
`from_dynamic_image`, which truncated the rgba value and shoved the rest
in a sbox. This box was then deleted by us.